Use stricter parsing when matching list blocks.
Previously I used a method that matched the block loosely and I seem to have intended to deal with it when parsing the list. That wasn't being done however.

I've switched the block matching to include the mandatory space after the list operator, but this means also matching the style directions.

closes #12