Refactor and update project to ES6.
The whole project has now be chopped down to smaller bits and ported to ES6 with the exception that it uses RequireJS modules, not ES6 ones. This is because Node.JS has full ES6 support apart from modules as of v6.0.0. The tests therefore run in a non-transpiled environment. Modules migration will happen eventually, but for now we use what we have.

The transpiled compatible versions are produced by webpack+babel. I have moved all built files to a dist directory. Thus, the general entrypoint is now dist/textile.js although the cli tries to use src/ if it detects a node version 6 or higher.

Tests were ported to AVA and cleaned up where applicable.

Eslintrc has been added to ensure code-consistency.

Polyfill for Array.isArray has been removed. This should be almost everywhere now, or user must provide their own (or stay on the 1.x version).

This is most of the work for issue #37.