Reworked blocks and HTML handling.
A lot of fixes made to how HTML and blocks are handled by the parser. Still holding on to the idea that mixing Textile and HTML can be sane, but only just. I believe we handle most cases gracefully, but I'm sure something will turn up one day to prove me wrong.

Fixes #36, #45