Modernized the code
Code now uses ESM import/export syntax. The package has not been set to type module as tape does not support that yet. The project is ready to be switched over but I am happy to wait for support in the test runner for now.

I have also cleaned up the tests so they now use template strings for multiline strings and arrow callback functions.

Moved babel config and browserlist to rc files.