Glyph conversions rewritten
The glyph conversions are run on all text nodes. This was a performance bottleneck so it has been rewritten for speed as well as given an accuracy overhaul.

The most notable change is that glyphs are no longer entity encoded by default. This seems like a silly default for JavaScript as well as fairly pointless if no other non-ascii entitles are being encoded. An option has been added to switch to the older behavior though.